Make sure the library data is stored for the proper user id
Previously the data of the reactive 'state' object is stored in the cache for the user id taken from 'props'. This can lead to storing false data for a user id, e.g. when the user navigates back and forth while the data is still fetching.

Now the fetched data is put in a separate object that is independent from the reactive 'state'.